+++ CodeWars/Validation/ValidBraces/Java/src/BraceCheckerTest.java
@@ -1,0 +1,19 @@
+
+import static org.junit.Assert.*;
+import org.junit.Test;
+
+public class BraceCheckerTest {
+
+    private BraceChecker checker = new BraceChecker();
+
+    @Test
+    public void testValid() {
+        assertEquals(true, checker.isValid("()"));
+    }
+
+    @Test
+    public void testInvalid() {
+        assertEquals(false, checker.isValid("[(])"));
+    }
+
+}